video: tegra: pva: Add nvhost_minimal_pm support
Add a driver that can do minimal pm support for
host1x CV modules, incase the bpmp is not enabled
in that platform. This driver is loaded before the
pva and dla driver get loaded.

update the NVIDIA-REVIEWERS file to add the
nvhost_minimal_pm dir

This changelist will be removed once bpmp is functional
with real silicon

JIRA PVA-381
JIRA PVA-481

Change-Id: I7a82b9b0da8fd4d2b3bb9bd27f14815377851500
Signed-off-by: Vinod G <vinodg@nvidia.com>
Reviewed-on: http://git-master/r/1483409
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>