crypto: Make tegra-cryptodev modulizable
The driver is to open an ioctl portal for user space to access
Tegra AES hardware, which should be modulizable according to
the original designer (Varun Wadekar <vwadekar@nvidia.com>).

So this patch just changes the driver from bool to tristate so
we can build it as a loadable module. Meanwhile, since it calls
the functions in the tegra-se-elp.c, this patch also adds this
dependency to make sure the module building be successful.

Bug 200305192

Change-Id: I9c7698e24d5ca10acbf62813e554da7c0ab6fcf6
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-on: http://git-master/r/1483469
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sachin Nikam <snikam@nvidia.com>