t19x: dts: Disable SMMU from PVA0 and PVA1
Due to simulator bugs PVA0 and PVA1 incorrectly do not drive smmu_ns
signal to SMMU. As a result, translation on PVA0 and PVA1 is failing
unless the overrides are enabled.

Further, due to model changes in SMMU, it is expected that the units
correctly drive streamid and smmu_ns signals and as such, the
translation fails with updated SMMU model.

This change temporarily disables SMMU from PVA0 and PVA1 to allow
updating the SMMU model.

Bug 200292136

Change-Id: I62d45547e211f5ae86118bf50887057b480f0aaa
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/1483778
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Pritesh Raithatha <praithatha@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Vinod Gopalakrishnakurup <vinodg@nvidia.com>