DT: soc: tegra: remove define WAKEUP_VIA_PM_IRQ_INTERRUPT_CONTROLLER
The support for the SOCs has moved to K4.4 and K4.9 which has
the macro defined WAKEUP_VIA_PM_IRQ_INTERRUPT_CONTROLLER. It
was not defined in k3.18.

As K3.18 is deprecated and hence this macros is not having any
meaning as it always defined. Remove this meaningless macro.

Change-Id: I1aeb32ad3f36fc3f2f9dedf59efc50c1a6fa5611
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/1483960
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit