DT: soc: t18x: pwm: Add slow parent as CLK_M
There is platform specific requirements to have PWM signal to
very lower frequency (1Hz). On this case, it is not possible
to achieve the PWM frequency with higher clock source.

Add support to have slow parent so that if usecase is to use
lower frequency then it can switch to slower parent.

bug 1927065

Change-Id: I66736f024e36d4eeffd7a9a3b675efcdf939ab9d
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/1484060
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: R Raj Kumar <rrajk@nvidia.com>
GVS: Gerrit_Virtual_Submit