DT: soc: t18x: Remove macro SDMMC_USE_CORE_REGULATOR_HANDLER
The macro SDMMC_USE_CORE_REGULATOR_HANDLER is defined to differentiate
the DT property for the SDMMC driver for K3.18 and K4.4. As support
moved to K4.4, the macro is default enabled for K4.4 and later version
of kernel. Support of K3.18 is already deprecated.

Hence it is not require to protect the DT property with macro and keep
the DT default for K44/later version.

Change-Id: Ie7f7efaf5b1aa248cfa789a264e1292fe2699bdc
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/1484647
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Pavan Kunapuli <pkunapuli@nvidia.com>
GVS: Gerrit_Virtual_Submit