dmaengine: tegra-adma: Fix system suspend handling
If ADMA is active when system suspend occurs, it is not suspended and
hence, the state and clock configuration is not handled correctly in
the suspend path.

Add proper system suspend and resume callbacks to ensure that the ADMA
is suspended and resumed if active when system suspend occurs.

Finally, make the system suspend callbacks for the ADMA late suspend
callbacks to ensure that the sound core has suspended any on-going
activity, including stopping the ADMA if active, before we attempt to
suspend the ADMA.

Bug 200275736

Change-Id: I7f24b5f21ad8ec76e5c129a1d384d11f6fc040a8
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: http://git-master/r/1488954
(cherry picked from commit 81dd03986638df9ca00fa166b317d127eca6ec37)
Reviewed-on: https://git-master.nvidia.com/r/1516276
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bibek Basu <bbasu@nvidia.com>