tegra: dc: improve accuracy of vblank timestamps
With the current appraoch, the monotonic
timestamp associated with the vblank event
gets updated only when the bottom half of the
dc-threaded interrupt handler gets scheduled. And
hence this results jitters in the timestamp
notified with this event. Add support to adjust
timestamp based on the current raster scanline
position.

Jira TDS-1352

Change-Id: Id6f8e6151c9d5227c5ea04d128478eb7156c3fb3
Signed-off-by: Arun Swain <arswain@nvidia.com>
Reviewed-on: http://git-master/r/1490884
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>