nvdisp: remove silicon/VDK check for HEAD rst
The section of code in the nvdisp driver that queries each
HEAD reset signal is currently protected by silicon and VDK
platform checks. This is entirely unnecessary - as long as bpmp
is running, the driver should be able to successfully retrieve
the reset handle. This change removes the extraneous platform
checks.

TDS-758

Change-Id: Icf155deb4feaf836cd30fdc7d4af344ab83dc7e1
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: http://git-master/r/1490975
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Aron Wong <awong@nvidia.com>
Reviewed-by: Ujwal Patel <ujwalp@nvidia.com>