ARM64: t210: jetson-e: Fix microphone bias routing
Jetson-Erista reports the following errors on boot:

 tegra-snd-t210ref-mobile-rt5639 sound: ASoC: Failed to add route x Mic Jack -> direct -> x micbias1
 tegra-snd-t210ref-mobile-rt5639 sound: ASoC: no sink widget found for x micbias1

The cause of the above errors comes from the fact that the rt5640
codec does not have a widget called 'micbias1' and so cannot be found.
The correct name for the widget is actually 'MICBIAS1' and this error
was introduced when switching the codec from the rt5639 which had a
widget called 'micbias1' to the rt5640 codec where the widget is now
called 'MICBIAS1'.

Please note that for the rt5640 codec 'MICBIAS1' is defined as a DAPM
'supply' and not a DAPM 'MICBIAS' as it was for rt5639. Therefore, the
order of the 'x MICBIAS1' and 'x Mic Jack' needs to be switched
otherwise the DAPM path will fail with the error ...

 'Connecting non-supply widget to supply widget is not supported'

Fix the name in the dts file and the ordering of the path to resolve
the audio routing for the microphone bias.

Bug 1926741

Change-Id: I9b6fa647d723d137a304b50e74185315052d3e8d
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: http://git-master/r/1493919
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ravindra Lokhande <rlokhande@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>