hdcp: fix for methods on T210
For T210 platforms on trusty, we need to enable the communication
between hdcp driver and hdcp TA so that the sessions can be launched
and cmac verification operations can be performed.
Earlier this was handled only for t18x, now modifications have been
made to make this happen for t210 too. Once the operations are launched
with the TZ, the returned values are sent to the hdcp ucode for cmac
validation of srm version.

Bug 200275348

Change-Id: Iae5559abba6647fe633b33843f52183a2f015bf1
Signed-off-by: Pranami Bhattacharya <pranamib@nvidia.com>
Reviewed-on: http://git-master/r/1494298
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Anand Prasad <anprasad@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Mitch Luban <mluban@nvidia.com>