t19x: cpuidle: Add cluster group support
This patch adds cluster group power state programming for cpuidle.
Deepest cluster group power state information is obtained from
device tree and is communicated to ccplex. Debugfs options are
provided for runtime modification of the state for testing.

This patch also fixes the cluster state re-programming
when a core is onlined or system exits sc7. The previous
code was restoring on incorrect cpu clusters rather
than the one that was onlined.

Change-Id: I3b17ca58c7644c09eafdada8ae2176fac2649e51
Signed-off-by: Krishna Sitaraman <ksitaraman@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1532618
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
Reviewed-by: Sai Gurrappadi <sgurrappadi@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Alexander Van Brunt <avanbrunt@nvidia.com>