bwmgr: provide default values
In case the SoC is not yet known to bwmgr (eg. when bringin up a new SoC
or kernel), provide default values so bwmgr can always be compiled in.
Ideally bwmgr should be split into a generic and per SoC files. The generic
part should either return an error on every call by a client in case of an
unknown SoC or provide default values.

Change-Id: I6790b25d0d98d1c09e458aed33e7f93001a5f1c8
Signed-off-by: Peter De Schrijver <pdeschrijver@nvidia.com>
Reviewed-on: http://git-master/r/1496279
GVS: Gerrit_Virtual_Submit
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
Reviewed-by: Alex Waterman <alexw@nvidia.com>
Reviewed-by: Timo Alho <talho@nvidia.com>