dc: fix lt-data child count check in of_dc
of_dc doesn't need to parse the LT tables from DT if either
the "lt-data" node doesn't exist, or if the node is empty.
With respect to the second condition, the panel selection redesign
changes that were merged a few months ago incorrectly removed the
negation of the of_get_child_count() check, which is causing the driver
to always ignore the "lt-data" node even if it exists. This change
fixes this issue.

Change-Id: Ib85ff42b637b2fee40ac66ce3c3d4135b72c8507
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: http://git-master/r/1498062
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
Reviewed-by: Mitch Luban <mluban@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ujwal Patel <ujwalp@nvidia.com>