video: tegra: host: Suspend unpowergatable devices
Currently suspend routines checks if there are pending references to
the device. However, since devices with power gating disabled take a
reference on the device during boot-up, the check in suspend fails
always.

This change modifies the check to take into account the case where
the device is not powergatable.

Bug 1936155

Change-Id: I71d4644b330a7d1f4033e7a4325daa4e087e92bb
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/1501308
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>