dp: stub out SSC option in dp_debug
SSC is a fixed clk PROD setting that is configured at boot,
and cannot be changed by SW. As such, the option that the
dp_debug framework exposes to toggle SSC for PLLDP is no
longer valid. To avoid breaking backwards-compatibility,
this patch makes the above option a no-op and prints an
informative warning message if the option is supplied by
the user, instead of completely removing it.

Bug 1940273

Change-Id: Iad75b57d2645c6c556dd5be582221781e9de062a
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: http://git-master/r/1502706
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Mitch Luban <mluban@nvidia.com>