video: tegra: host: Abort suspend if an interrupt waiter exists
During suspend we disable interrupts, but that can fail if there is a
waiter attached to a syncpt threshold. However, we currently don't
propagate any error code, so suspend "succeeds" whether or not there
is an error.

Propagate the error code and abort suspend if disabling interrupts
fails.

Jira HOSTX-228

Change-Id: I0b3b3dea7f54430fe110e0bde74eacff30619224
Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
Reviewed-on: http://git-master/r/1503843
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Arto Merilainen <amerilainen@nvidia.com>