video: tegra: host: Drop extra runtime PM reference during suspend
When a device is marked as non-powergateable, we take an extra runtime
PM reference to it during initialization. Due to this extra reference,
when doing system resume, the device gets automatically resumed as
well, which is not intended. Therefore drop the reference when
preparing for suspend and retake it using get_noresume when coming
out of suspend.

Bug 1936155

Change-Id: I4851318de9650c650f3ea60f16aa6fbec8f53025
Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
Reviewed-on: http://git-master/r/1504808
Reviewed-by: Deepak Nibade <dnibade@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
Reviewed-by: Arto Merilainen <amerilainen@nvidia.com>