video: tegra: host: pva: Fix output status writes
Currently code assumes that the value passed for status write action
is written in case where the task was successful. However, this is
incorrect: In case the task succeeds, PVA writes 0 to the status
structure. In other cases PVA writes the error code defined in the
action.

Since status write was using an incorrect code, PVA always wrote
value 0 regardless of the real task status. This patch changes code
to use the correct error code while writing the actions.

JIRA PVA-496

Change-Id: Ifc06502b1582e6dd3794e2517a9ddc8e10642ac9
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/1505671
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>