arm64: configs: tegra: Revert TRUSTY to =y
The display HDMIHDCP driver is defined as a boolean type driver
while it covertly depends on the trusty driver by calling one of
the ote functions. As we cannot change the HDCP driver to =m in
a short time. This forces us to revert the trusty back to =y so
as to let the HDCP driver get compiled correctly.

boot.img is increased by 59392 bytes

Bug 1938823

Change-Id: I3156216e1e300dcdae15124b814685f042a7e0c1
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-on: http://git-master/r/1505889
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>