video: tegra: nvmap: Cache clean generic carveout allocations
When generic carveout is allocated as CMA memory, it can
be accessed by kernel. DMA alloc guarantees that the allocations
are zeroed out.

Since, nvmap has to guarantee that
the allocations from generic carveout are memset to 0 and
are written back to DRAM, the cache clean is neccessary.

Bug 1942379

Change-Id: Iba8ee5ec33edbfcf471c9ea33fd2bf806814cbb8
Signed-off-by: Sri Krishna chowdary <schowdary@nvidia.com>
Reviewed-on: http://git-master/r/1506859
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sachin Nikam <snikam@nvidia.com>