ASoC: tegra-alt: Fix Orca Viper I2S master setup
Orca Viper has 4 I2S channels, where one channel is configured to
drive the bit-clock and frame-sync for all channels. When one of the
slave channels is initialised the clocks for master I2S channel need
to be enabled so that the bit-clock is on. Otherwise the resetting of
the slave channel will fail because the bit-clock is not running.

When configuring the master I2S channel, ensure that the master bit
for the I2S channel is set at this point, otherwise the bit clock will
not turn on when turning on the I2S master channel clocks and this may
cause the slave channel soft reset to fail. Previously, we were only
setting the master bit for the I2S master when the I2S master itself
is started.

Bug 200314397

Change-Id: I188abb79cd371c1573a7e4b2b7ec866e40d98982
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master/r/1510908
(cherry picked from commit 2ba5d517cbf3919f01b4753ebda2be58430c5842)
Reviewed-on: https://git-master/r/1512531
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bibek Basu <bbasu@nvidia.com>