ASoC: tegra-alt: Fix Orca Viper bclk ratio setting
The bit-clock ratio is not being set-up per the device-tree
configuration and so the ratio defaults to 2, which means the bit-clock
is running twice the speed it should be. Fix this by programming the
bit-clock ratio as defined in device-tree.

Bug 200286098

Change-Id: I91027b2a70e44cf3edb83c1d1c91bda826bafcbc
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master/r/1510937
(cherry picked from commit 88a36366393ace6ed088bb6667538127230a8323)
Reviewed-on: https://git-master/r/1512532
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bibek Basu <bbasu@nvidia.com>