dc: add hooks to program clk mux on T19x FPGA
On T19x FPGA, all display-related clks are fixed. As such,
any special clocking requirements that are needed for a given
output mode are handled automatically by an SOR clk mux that SW
needs to program accordingly. This patch defines a generic
SOR hook for this purpose, and calls it during the HDMI and DP
controller enable sequences.

TDS-2396

Change-Id: Id1f6f88a2eaf80e191ec354d3853fa8bd9655b06
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1511433
Reviewed-by: Aron Wong <awong@nvidia.com>
Reviewed-by: Pradeepan Patra <ppatra@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ujwal Patel <ujwalp@nvidia.com>
Reviewed-by: Santosh Galma <galmar@nvidia.com>