scripts: update checkpatch.pl with latest K4.9 snapshot
This contains a fix for an issue seen when a new file is added.
The issue reported is:

Use of uninitialized value $root in concatenation (.) or
string at /tmp/tmp3tca6a line 767.

Also add const_structs.checkpatch file which is added for K4.9,
which will fix the following issue reported in ACV:

No structs that should be const will be found - file
'/tmp/const_structs.checkpatch': No such file or directory

Bug 1928069

Change-Id: I72b8a7b0d3e5c0f3d68258ce68e9961987275bb3
Signed-off-by: Bharat Nihalani <bnihalani@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1515451
Reviewed-by: Jani Uusi-Rantala <juusirantala@nvidia.com>