firmware: tegra: remove tegra_bpmp_send()
tegra_bpmp_send() was designed as a "one way" message to bpmp firmware
if the calling driver want to avoid waiting for the response (because it
does not care about the response). At the time of design, it was thought
that this sort of functionality will be needed more often than it is
used today. Implementation of this API requires somewhat unusual
semantics at the handler side (=bpmp firmware) where it touches bits
that are otherwise modified by the kernel.

Therefore, remove this function from driver.

Change-Id: I2b88898e338b7d3c56a145cc00d9838a487cda92
Signed-off-by: Sivaram Nair <sivaramn@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1516526
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Matthew Longnecker <mlongnecker@nvidia.com>
GVS: Gerrit_Virtual_Submit