firmware: tegra: support single atomic channel
Currently, we assume that there are as many atomic channels as the
number of available CPUs (and return error when this assumption is
broken). In order to support reducing the number of atomic channels to
1, change the tegra_bpmp_send_receive_atomic() to use a spin lock for
protecting the (single) atomic channel.

Additionally, since this might create a non-contiguous channel numbering
on existing platforms, use a bit mask to represent enabled channels.

Bug 1956912

Change-Id: Ifa4390ff47848071c291c54041412f271f4e34a8
Signed-off-by: Sivaram Nair <sivaramn@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1516527
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>