dc: enable HDMI HDCP init on FPGA
HDCP init was intially stubbed out in the HDMI driver on FPGA.
At that time, DDC/I2C reads hadn't been verified yet, which
meant that the DDC bus was omitted in DT. This caused issues
since the HDCP driver expects a valid DDC bus to be passed in
to correctly register its I2C adapter.

Now that DDC/I2C reads are verified, this change goes ahead
and enables HDMI HDCP init on FPGA. Additionally, this resolves
a memory access crash that's seen on TOT whenever a hotplug event
is triggered. Currently, the FPGA platform check is only applied to
HDCP init, but not to any of the other places in the HDMI driver
which send plug/unplug notifications to the HDCP driver. This is
very problematic since the plug/unplug callbacks will attempt to
access an internal HDCP struct which hasn't been allocated
at all.

TDS-2400

Change-Id: Ib1a234911dcd0286bd1b8aa3998384c618d0563d
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1517674
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Pranami Bhattacharya <pranamib@nvidia.com>
Reviewed-by: Ujwal Patel <ujwalp@nvidia.com>